Se cero carpeta para código de google maps por fuera de la app y se pasan los archivos que estaban como componentes y en utils
