--- front/.env
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_GOOGLE_MAPS_API_KEY = ''
