+++ front/.gitignore
@@ -26,8 +26,9 @@ yarn-debug.log*
 yarn-error.log*
 
 # local env files
-.env*.local
 .env
+.env*.local
+
 
 # vercel
 .vercel
